Harden spelling helper interfaces
Group refresh policy and transport state into immutable values so the
shared spelling helper remains within the estate's code-health limits.
Split refresh tests into focused modules to keep every source file below
the repository's 400-line limit while preserving the full contract suite.
