+++ scripts/tests/conftest.py
@@ -1,0 +1,23 @@
+"""Shared fixtures for spelling rollout tests."""
+
+from __future__ import annotations
+
+import importlib
+import types
+from pathlib import Path
+
+import pytest
+
+SCRIPT_DIRECTORY = Path(__file__).resolve().parents[1]
+
+
+@pytest.fixture(name="rollout_modules")
+def rollout_modules_fixture(
+    monkeypatch: pytest.MonkeyPatch,
+) -> tuple[types.ModuleType, types.ModuleType, types.ModuleType]:
+    """Import scripts through the top-level paths used at runtime."""
+    monkeypatch.syspath_prepend(str(SCRIPT_DIRECTORY))
+    names = ("typos_rollout_cache", "typos_rollout", "generate_typos_config")
+    importlib.invalidate_caches()
+    cache, rollout, generator = (importlib.import_module(name) for name in names)
+    return cache, rollout, generator
