+++ scripts/tests/typos_rollout_test_support.py
@@ -1,0 +1,11 @@
+"""Shared data builders for spelling rollout tests."""
+
+
+def dictionary_text(stem: str = "organ") -> str:
+    """Return a minimal valid shared-dictionary document."""
+    return (
+        'schema = 1\n\n[oxford]\nstems = ["'
+        + stem
+        + '"]\n\n[words]\naccepted = []\n\n[words.corrections]\n\n'
+        + "[patterns]\nignore = []\n\n[files]\nexclude = []\n"
+    )
